+++ qom/ui/widgets/LooperWidget.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.ui.widgets.LooperWidget'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-08-19'
-__updated__ = '2021-08-28'
+__updated__ = '2021-08-30'
 
 # dependencies
 from PyQt5 import QtCore, QtGui, QtWidgets
@@ -18,7 +18,7 @@ import os
 # qom modules
 from ...loopers import XLooper, XYLooper, XYZLooper
 from ...utils import default_looper_func_names, get_looper_func
-from . import BaseWidget
+from .BaseWidget import BaseWidget
 
 # module logger
 logger = logging.getLogger(__name__)