+++ qom/ui/widgets/SystemWidget.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.ui.widgets.SystemWidget'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-01-21'
-__updated__ = '2021-08-28'
+__updated__ = '2021-08-30'
 
 # dependencies
 from PyQt5 import QtCore, QtGui, QtWidgets
@@ -18,7 +18,8 @@ import os
 import re
 
 # qom modules
-from . import BaseWidget, ParamWidget
+from .BaseWidget import BaseWidget
+from .ParamWidget import ParamWidget
 
 # module logger
 logger = logging.getLogger(__name__)